Separated elements views from the others
