+++ src/main/kotlin/automaton/constructor/controller/AutomatonGraphController.kt
@@ -7,6 +7,8 @@ import automaton.constructor.model.element.*
 import automaton.constructor.utils.*
 import automaton.constructor.view.*
 import automaton.constructor.view.automaton.AutomatonGraphView
+import automaton.constructor.view.elements.vertex.AutomatonVertexView
+import automaton.constructor.view.elements.transition.TransitionView
 import javafx.geometry.Point2D
 import javafx.scene.control.ContextMenu
 import javafx.scene.input.KeyCode
