+++ src/main/kotlin/automaton/constructor/view/elements/transition/BasicTransitionView.kt
@@ -1,10 +1,11 @@
-package automaton.constructor.view
+package automaton.constructor.view.elements.transition
 
 import automaton.constructor.model.element.Transition
 import automaton.constructor.utils.I18N
 import automaton.constructor.utils.Setting
 import automaton.constructor.utils.SettingGroup
 import automaton.constructor.utils.createUnmodifiableSettingControl
+import automaton.constructor.view.AutomatonElementView
 import tornadofx.toProperty
 
 open class BasicTransitionView(val transition: Transition): AutomatonElementView(transition) {
