+++ src/main/kotlin/automaton/constructor/view/elements/transition/TableTransitionView.kt
@@ -1,5 +1,6 @@
-package automaton.constructor.view
+package automaton.constructor.view.elements.transition
 
 import automaton.constructor.model.element.Transition
+import automaton.constructor.view.elements.transition.BasicTransitionView
 
 open class TableTransitionView(transition: Transition): BasicTransitionView(transition)
