+++ src/main/kotlin/automaton/constructor/view/elements/vertex/AutomatonBasicVertexView.kt
@@ -1,17 +1,12 @@
-package automaton.constructor.view
+package automaton.constructor.view.elements.vertex
 
 import automaton.constructor.model.element.AutomatonVertex
-import automaton.constructor.model.element.BuildingBlock
-import automaton.constructor.model.module.hasProblems
-import automaton.constructor.model.module.hasProblemsBinding
 import automaton.constructor.utils.I18N
 import automaton.constructor.utils.Setting
 import automaton.constructor.utils.SettingGroup
-import javafx.beans.property.SimpleIntegerProperty
-import javafx.beans.property.SimpleStringProperty
+import automaton.constructor.view.AutomatonElementView
 import javafx.scene.control.CheckBox
 import javafx.scene.control.TextField
-import javafx.scene.paint.Color
 import tornadofx.*
 
 open class AutomatonBasicVertexView(val vertex: AutomatonVertex) : AutomatonElementView(vertex) {
